# Conflicts:
#	app/src/main/java/com/blipblipcode/distribuidoraayl/core/local/entities/openFactura/reportSale/SaleDataEntity.kt
#	app/src/main/java/com/blipblipcode/distribuidoraayl/core/local/room/DataBaseApp.kt
#	app/src/main/java/com/blipblipcode/distribuidoraayl/core/local/room/DatabaseModule.kt
#	app/src/main/java/com/blipblipcode/distribuidoraayl/data/repositiry/of/OpenFacturaRepository.kt
#	app/src/main/java/com/blipblipcode/distribuidoraayl/ui/utils/ScopeFunction.kt
#	app/src/main/res/values/strings.xml
#	gradle/libs.versions.toml

+++ app/src/main/java/com/blipblipcode/distribuidoraayl/data/dto/of/EmissionResponseDto.kt
@@ -6,7 +6,7 @@ data class EmissionResponseDto (
     @SerializedName("TOKEN")
     val token: String,
     @SerializedName("FOLIO")
-    val number: Int,
+    val number: Long,
     @SerializedName("RESOLUCION")
     val resolution: ResolutionDto,
     @SerializedName("TIMBRE")
